+++ CHANGELOG.md
@@ -4,10 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.7.0] - 2020-12-23
+### Changed
+- added minimum supported Android Build Tools version to 4.1.0
+- update minimum supported Dokka version to 1.4.10
+
 ## [0.6.0] - 2020-07-13
 ### Changed
 - added support for Android Build Tools 4.0.0
-- udpate supported Dokka version
+- update supported Dokka version
 
 ## [0.5.2] - 2020-04-09
 ### Changed